requirements.txt
# Discord
discord.py>=2.3.0

# HTTP Client
httpx>=0.25.0

# Google Sheets
gspread>=5.12.0
google-auth>=2.23.0

# Image Processing
Pillow>=10.0.0

# Data Validation & Config
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Code Quality
black>=23.0.0
isort>=5.12.0
mypy>=1.7.0